+++ scripts/image_quality_analysis.py
